on_demand_project/requirements.txt
django
psycopg2>=2.7
djangorestframework
djangorestframework-simplejwt
drf-yasg
python-dotenv

